# Add optimised max_element implementation as a specific case of nth_element with beam = 1