--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Optimised special case for nth_element when decoding with beam size of 1.
+
 ## [1.12.0] - 2023-02-20
 
 ### Added