@XapaJIaMnu 2023-04-17
Apologies for the late reply @hieuhoang , I have some benchmarks, finally.
I show a tiny11.tied.w configuration from WMT a few years ago, tested on 15.5k WMT sentences from the last many years (avg BLEU 36.{2,4,7} depending on configuration) with and without shortlist, using upstream fbgemm/intgemm/fp32 cpu backends). Here's the results:

tl;dr the beam1 code is 1-5 seconds faster depending on the test case. The bigger the output layer, the larger the difference.

|                               | Beam1Opt |  Master |
|:-----------------------------:|:--------:|:-------:|
| AVX512, fbgemm, shortlist     | 94.02s   | 96.47s  |
| AVX512, fbgemm, no shortlist  | 182.80s  | 188.05s |
| AVX2,fbgemm, shortlist        | 114.26s  | 115.75s |
| AVX2, fbgemm, no shortlist    | 195.78s  | 200.80s |
| AVX512, intgemm, shortlist    | 106.20s  | 108.71s |
| AVX512, intgemm, no shortlist | 194.08s  | 200.36s |
| AVX2,intgemm, shortlist       | 119.08s  | 120.79s |
| AVX2, intgemm, no shortlist   | 200.11s  | 205.18s |
| AVX512, shortlist             | 118.29s  | 120.85s |
| AVX512, no shortlist          | 209.82s  | 216.63s |
| AVX2, shortlist               | 135.16s  | 136.96s |
| AVX2, no shortlist            | 215.94s  | 221.30s |


To download the models and test for yourself, please get this tarball https://nbogoychev.com/files/speedtest.tar.gz