@hieuhoang 2023-04-17
I have no objections to approving the PR. Nick's results show a slight improvement for his model. My result, below show hardly any change. Inching forward.

<html xmlns:v="urn:schemas-microsoft-com:vml"
xmlns:o="urn:schemas-microsoft-com:office:office"
xmlns:x="urn:schemas-microsoft-com:office:excel"
xmlns="http://www.w3.org/TR/REC-html40">

<head>

<meta name=ProgId content=Excel.Sheet>
<meta name=Generator content="Microsoft Excel 15">
<link id=Main-File rel=Main-File
href="file:///C:/Users/hihoan/AppData/Local/Temp/msohtmlclip1/01/clip.htm">
<link rel=File-List
href="file:///C:/Users/hihoan/AppData/Local/Temp/msohtmlclip1/01/clip_filelist.xml">
<style>
<!--table
	{mso-displayed-decimal-separator:"\.";
	mso-displayed-thousand-separator:"\,";}
@page
	{margin:.75in .7in .75in .7in;
	mso-header-margin:.3in;
	mso-footer-margin:.3in;}
tr
	{mso-height-source:auto;}
col
	{mso-width-source:auto;}
br
	{mso-data-placement:same-cell;}
td
	{padding-top:1px;
	padding-right:1px;
	padding-left:1px;
	mso-ignore:padding;
	color:black;
	font-size:11.0pt;
	font-weight:400;
	font-style:normal;
	text-decoration:none;
	font-family:Calibri, sans-serif;
	mso-font-charset:0;
	mso-number-format:General;
	text-align:general;
	vertical-align:bottom;
	border:none;
	mso-background-source:auto;
	mso-pattern:auto;
	mso-protection:locked visible;
	white-space:nowrap;
	mso-rotate:0;}
.xl66
	{font-weight:700;}
-->
</style>
</head>

<body link="#0563C1" vlink="#954F72">



  | master | nick's max_element
-- | -- | --
Big Machine | 52.85 | 51.74
  | 53 | 52.49
  | 51.11 | 52.4
+bin model | 154.16 | 154.2
  | 150.89 | 153.32
  | 150.77 | 155.23
base+LSH | 135.02 | 137.37
  | 135.27 | 136.12
  | 133.43 | 135.39
bin model+LSH | 52.79 | 53.34
  | 53.08 | 52.58
  | 52.89 | 54.04
  |   |  
bin model, beam2 | 36.12 | 35.51
bin model, beam3 | 40.43 | 40.73
bin model, beam4 | 62.01 | 62.77
  |   |  



</body>

</html>